Refactor reverseComplement to SAMRecord and make default copy attributes
