+++ src/main/java/htsjdk/samtools/SAMRecord.java
@@ -26,17 +26,12 @@ package htsjdk.samtools;
 
 import htsjdk.samtools.util.CoordMath;
 import htsjdk.samtools.util.Locatable;
+import htsjdk.samtools.util.SequenceUtil;
 import htsjdk.samtools.util.StringUtil;
 
 import java.io.Serializable;
 import java.lang.reflect.Array;
-import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.Collections;
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
-import java.util.Set;
+import java.util.*;
 
 
 /**
@@ -160,6 +155,16 @@ public class SAMRecord implements Cloneable, Locatable, Serializable {
      * abs(insertSize) must be <= this
      */
     public static final int MAX_INSERT_SIZE = 1<<29;
+
+    /**
+     * Tags that are known to need the reverse complement if the read is reverse complemented.
+     */
+    public static List<String> TAGS_TO_REVERSE_COMPLEMENT = Arrays.asList(SAMTag.E2.name(), SAMTag.SQ.name());
+
+    /**
+     * Tags that are known to need the reverse if the read is reverse complemented.
+     */
+    public static List<String> TAGS_TO_REVERSE = Arrays.asList(SAMTag.OQ.name(), SAMTag.U2.name());
 
     private String mReadName = null;
     private byte[] mReadBases = NULL_SEQUENCE;
@@ -2249,5 +2254,127 @@ public class SAMRecord implements Cloneable, Locatable, Serializable {
         if (this.transientAttributes != null) return this.transientAttributes.remove(key);
         else return null;
     }
-}
 
+    /**
+     * Reverse-complement bases and reverse quality scores along with known optional attributes that
+     * need the same treatment. Changes made after making a copy of the bases, qualities,
+     * and attributes. If in-place update is needed use {@link #reverseComplement(SAMRecord, boolean)}.
+     * See {@link #TAGS_TO_REVERSE_COMPLEMENT} {@link #TAGS_TO_REVERSE}
+     * for the default set of tags that are handled.
+     */
+    public static void reverseComplement(final SAMRecord rec) {
+        reverseComplement(rec, false);
+    }
+
+    /**
+     * Reverse-complement bases and reverse quality scores along with known optional attributes that
+     * need the same treatment. Optionally makes a copy of the bases, qualities or attributes instead
+     * of altering them in-place. See {@link #TAGS_TO_REVERSE_COMPLEMENT} {@link #TAGS_TO_REVERSE}
+     * for the default set of tags that are handled.
+     *
+     * @param rec Record to reverse complement.
+     * @param inplace Setting this to false will clone all attributes, bases and qualities before changing the values.
+     */
+    public static void reverseComplement(final SAMRecord rec, boolean inplace) {
+        reverseComplement(rec, TAGS_TO_REVERSE_COMPLEMENT, TAGS_TO_REVERSE, inplace);
+    }
+
+    /**
+     * Reverse complement bases and reverse quality scores. In addition reverse complement any
+     * non-null attributes specified by tagsToRevcomp and reverse and non-null attributes
+     * specified by tagsToReverse.
+     */
+    public static void reverseComplement(final SAMRecord rec, final Collection<String> tagsToRevcomp, final Collection<String> tagsToReverse, boolean inplace) {
+        final byte[] readBases = inplace ? rec.getReadBases() : rec.getReadBases().clone();
+        SequenceUtil.reverseComplement(readBases);
+        rec.setReadBases(readBases);
+        final byte qualities[] = inplace ? rec.getBaseQualities() : rec.getBaseQualities().clone();
+        reverseArray(qualities);
+        rec.setBaseQualities(qualities);
+
+        // Deal with tags that need to be reverse complemented
+        if (tagsToRevcomp != null) {
+            for (final String tag: tagsToRevcomp) {
+                Object value = rec.getAttribute(tag);
+                if (value != null) {
+                    if (value instanceof byte[]) {
+                        value = inplace ? value : ((byte[]) value).clone();
+                        SequenceUtil.reverseComplement((byte[]) value);
+                    }
+                    else if (value instanceof String) {
+                        //SequenceUtil.reverseComplement is in-place for bytes but copies Strings since they are immutable.
+                        value = SequenceUtil.reverseComplement((String) value);
+                    }
+                    else throw new UnsupportedOperationException("Don't know how to reverse complement: " + value);
+                    rec.setAttribute(tag, value);
+                }
+            }
+        }
+
+        // Deal with tags that needed to just be reversed
+        if (tagsToReverse != null) {
+            for (final String tag : tagsToReverse) {
+                Object value = rec.getAttribute(tag);
+                if (value != null) {
+                    if (value instanceof String) {
+                        value = StringUtil.reverseString((String) value);
+                    }
+                    else if (value.getClass().isArray()) {
+                        if (value instanceof byte[]) {
+                            value = inplace ? value : ((byte[]) value).clone();
+                            reverseArray((byte[]) value);
+                        }
+                        else if (value instanceof short[]) {
+                            value = inplace ? value : ((short[]) value).clone();
+                            reverseArray((short[]) value);
+                        }
+                        else if (value instanceof int[]) {
+                            value = inplace ? value : ((int[]) value).clone();
+                            reverseArray((int[]) value);
+                        }
+                        else if (value instanceof float[]) {
+                            value = inplace ? value : ((float[]) value).clone();
+                            reverseArray((float[]) value);
+                        }
+                        else throw new UnsupportedOperationException("Reversing array attribute of type " + value.getClass().getComponentType() + " not supported.");
+                    }
+                    else throw new UnsupportedOperationException("Don't know how to reverse: " + value);
+
+                    rec.setAttribute(tag, value);
+                }
+            }
+        }
+    }
+
+    private static void reverseArray(final byte[] array) {
+        for (int i=0, j=array.length-1; i<j; ++i, --j) {
+            final byte tmp = array[i];
+            array[i] = array[j];
+            array[j] = tmp;
+        }
+    }
+
+    private static void reverseArray(final short[] array) {
+        for (int i=0, j=array.length-1; i<j; ++i, --j) {
+            final short tmp = array[i];
+            array[i] = array[j];
+            array[j] = tmp;
+        }
+    }
+
+    private static void reverseArray(final int[] array) {
+        for (int i=0, j=array.length-1; i<j; ++i, --j) {
+            final int tmp = array[i];
+            array[i] = array[j];
+            array[j] = tmp;
+        }
+    }
+
+    private static void reverseArray(final float[] array) {
+        for (int i=0, j=array.length-1; i<j; ++i, --j) {
+            final float tmp = array[i];
+            array[i] = array[j];
+            array[j] = tmp;
+        }
+    }
+}
